Split models into separate files, rename tables and columns for clarity
